+++ flask_dapr/__init__.py
@@ -5,7 +5,7 @@ Copyright (c) Microsoft Corporation.
 Licensed under the MIT License.
 """
 
-from flask_dapr.actor import DaprActor
+from .actor.actor import DaprActor
 from .version import __version__
 
 __all__ = ['DaprActor', '__version__', ]
