+++ mypy.ini
@@ -10,8 +10,5 @@ files =
     dapr/clients/**/*.py,
     dapr/conf/**/*.py,
     dapr/serializers/**/*.py,
-    flask_dapr/**/*.py,
+    flask_dapr/actor/*.py,
     examples/demo_actor/**/*.py
-
-[mypy-flask_dapr.setup.py]
-ignore_errors = True
